perf(web): isolate currentTime media-session sync into a leaf
useMediaSession() ran at the App root and subscribed to currentTime, which
the audio engine writes every 250ms — so the entire inline App tree
reconciled 4x/sec during playback.

Extract the two currentTime-dependent effects (OS overlay position state +
throttled 1Hz playback-state IPC) into an isolated <MediaSessionSync/> leaf
that subscribes to currentTime itself and renders null, mirroring the
existing TimeDisplay pattern. useMediaSession keeps the action handlers,
metadata, and playbackState (none of which depend on currentTime), so App
no longer re-renders on every time tick.
