+++ apps/web/src/hooks/useMediaSession.ts
@@ -1,4 +1,4 @@
-import { useEffect, useRef } from 'react';
+import { useEffect } from 'react';
 import { usePlaybackStore } from '@/stores/usePlaybackStore';
 import { IS_ELECTRON } from '@/lib/platform';
 
@@ -6,21 +6,20 @@ import { IS_ELECTRON } from '@/lib/platform';
  * Integrates with:
  * 1. navigator.mediaSession — for OS media overlay (Windows media popup, macOS Now Playing)
  * 2. Electron IPC — listens for media:command from main process (global shortcuts)
- * 3. Sends playback state back to main process for tray/taskbar updates
+ *
+ * This hook intentionally does NOT subscribe to currentTime — that high-frequency
+ * value (written every 250ms) is owned by the isolated <MediaSessionSync/> leaf so
+ * the host (root App) does not re-render 4x/sec. Mount <MediaSessionSync/> once
+ * alongside this hook.
  */
 export function useMediaSession() {
   const currentTrack = usePlaybackStore(s => s.currentTrack);
   const isPlaying = usePlaybackStore(s => s.isPlaying);
-  const currentTime = usePlaybackStore(s => s.currentTime);
-  const duration = usePlaybackStore(s => s.duration);
   const togglePlay = usePlaybackStore(s => s.togglePlay);
   const next = usePlaybackStore(s => s.next);
   const previous = usePlaybackStore(s => s.previous);
   const stop = usePlaybackStore(s => s.stop);
   const seek = usePlaybackStore(s => s.seek);
-
-  // Throttle state updates to main process (every 1 second)
-  const lastUpdateRef = useRef(0);
 
   // Set up Media Session API action handlers
   useEffect(() => {
@@ -116,20 +115,6 @@ export function useMediaSession() {
     navigator.mediaSession.playbackState = isPlaying ? 'playing' : 'paused';
   }, [isPlaying]);
 
-  // Update position state for seek bar in OS media overlay
-  useEffect(() => {
-    if (!('mediaSession' in navigator) || !duration || !isFinite(duration)) return;
-    try {
-      navigator.mediaSession.setPositionState({
-        duration,
-        playbackRate: 1,
-        position: Math.min(currentTime, duration),
-      });
-    } catch {
-      // Some browsers don't support setPositionState
-    }
-  }, [currentTime, duration]);
-
   // Listen for media commands from Electron main process
   useEffect(() => {
     if (!IS_ELECTRON) return;
@@ -153,28 +138,4 @@ export function useMediaSession() {
 
     return unsub;
   }, [togglePlay, next, previous, stop]);
-
-  // Send playback state to main process (throttled)
-  useEffect(() => {
-    if (!IS_ELECTRON) return;
-
-    const now = Date.now();
-    if (now - lastUpdateRef.current < 1000 && lastUpdateRef.current > 0) return;
-    lastUpdateRef.current = now;
-
-    if (!currentTrack) {
-      window.electronAPI.media.clearState();
-      return;
-    }
-
-    window.electronAPI.media.sendPlaybackState({
-      isPlaying,
-      title: currentTrack.title,
-      artist: currentTrack.artist,
-      album: currentTrack.album,
-      duration,
-      currentTime,
-      albumArt: currentTrack.albumArt ?? null,
-    });
-  }, [currentTrack, isPlaying, currentTime, duration]);
 }
